+++ docs/docs/changelog.md
@@ -16,9 +16,12 @@ Types of changes:
 - `Security` in case of vulnerabilities.
 
 ## [Unreleased] - {ISO8601 date}
+
+## [0.20.0] - 2021-04-15
 ### Added
 - `read_fit() now returns pool length records when the `pool_lenghts=True` argument is passed.
 - `read_fit()` now return raw FIT messages when a `raw_message=True` argument is passed.
+- `read_fit()` can now return metadata and device information.
 
 ## [0.19.0] - 2021-04-06
 ### Added